- notifications - settings - notifications popover

+++ app/Http/Controllers/FollowController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use App\Notifications\UserFollowedNotification;
 use Illuminate\Http\Request;
 
 class FollowController extends Controller
@@ -18,8 +19,17 @@ class FollowController extends Controller
 
         if ($currentUser->isFollowing($userToFollow)) {
             $currentUser->following()->detach($userToFollow->id);
+
+            // Delete the follow notification
+            $userToFollow->notifications()
+                ->where('type', UserFollowedNotification::class)
+                ->where('data->actor_id', $currentUser->id)
+                ->delete();
         } else {
             $currentUser->following()->attach($userToFollow->id);
+
+            // Send notification
+            $userToFollow->notify(new UserFollowedNotification($currentUser));
         }
 
         return back();
